requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
google-generativeai
faiss-cpu
pymupdf
streamlit
python-dotenv